+++ th902-main/src/cn/s3bit/th902/gamecontents/components/player/ReimuBullet1.java
@@ -148,7 +148,7 @@ public class ReimuBullet1 extends Component implements IJudgeCallback {
 	}
 
 	@Override
-	public int getDamage() {
+	public float getDamage() {
 		return mDamage;
 	}
